chore: DWH cleanup
